Add MXNet backend (#83)
* Add placeholder MXNet backend

Add a placeholder file for the MXNet backend and partially fill it in.

In addition also add mxnet to packages to pip install and add
mxnet_backend to the list of backends to test.

* Add MXNet backend stack() and concatenate()

Other methods are added as well, but these are the two that were the
most important

* Add placeholder for outer()

The current outer() function is highly suboptimal. However, until a
better solution can be implimented in a more MXNet-centric style it will
at least pass the tests. The Pull Request in which this commit exists
should _NOT_ be merged in until outer() is updated with a better
solution.

* Use MXNet-centric implimentation of outer()

Use the linear algebra functions of MXNet to provide a more clean
implimenation of the outer product method, outer()

* Add mxnet_backend try block

* Add default Jupyter notebook example using MXNet backend

* Use list indexing instead of extended iterable unpacking

Extended iterable unpacking wasn't added until PEP 3132 and so isn't in
Python 2, which is causing the Python 2.7 tests to fail. While this is
annoying, it is probably worth still supporting Python 2.7 until the
official Python 2 end of life date.

c.f. https://www.python.org/dev/peps/pep-3132/

* Use Conda environment.yml for Binder

To support the more complex systems of multiple backends a Conda
environment.yml file is used instead of a pip requirements.txt file for
use with Binder.

c.f.
https://mybinder.readthedocs.io/en/latest/sample_repos.html#conda-environment-with-environment-yml

* Have simple_broadcast() return a MXNet NDArray

Additionally make a correction to outer().
nd.broadcast_mul() is the correct function to use in this method.

* Add docstrings

Add Google Style docstrings that should be compatible with Sphinx under
certain themes, such as the Sphinx theme created by Read The Docs. For
an example of what this would look like, see PyTorch's docs.

c.f.:
  http://sphinxcontrib-napoleon.readthedocs.io/en/latest/example_google.html#example-google
  http://pytorch.org/docs/master/_modules/torch/distributions/bernoulli.html#Bernoulli