Benchmark runOnePoint()
Benchmark runOnePoint() so that a fit is actually done and there will be
variation in timing.

At the moment only the NumPy and PyTorch backends are benchmarked in the
test function as there are scaling problems with TensorFlow and the
MXNet optimizer has not been completed.

The source that is used is just repeating values to ensure
reproducibility in the CI. Poisson variations with a fixed seed could
also work, but preliminary testing shows that the NumPy optimizer, which
does not take advantage of auto differentiation, is not always able to
complete the fit at high bin values which would result in a failing test.
Of interest is that PyTorch is able to do so.