Add benchmarks using pytest-benchmark (#92)
* Add initial benchmarks using pytest-benchmark

Add some proof of concept benchmakrs using the features of
pytest-benchmark
c.f. https://github.com/ionelmc/pytest-benchmark

* [Temporary] Parameterize only number of bins

The parameterization of the backends is causing changes in the testing
environment which cause the rest of the tests to fail. For the time
being turn off backends and only parameterize the number of bins.

Additionally beautify test_import.py

* Reset pyhf.tensorlib backend each test

After each test where the pyhf backend is changed reset the backend to
the default value of pyhf.tensorlib

Use ids with the backends to more clearly indicate which backend is
being used by id-ing it by its name

* Add bin ids

Add bin ids for cleaner labeling

Install pytest-benchmark with [histogram] option to also install pygal
so that the --benchmark-histogram option works

* Expand range to match bins and vary bin content

Have the range of the histogrm expand as more bins are added

Additionally, instead of adding the same bin content for every bin, have
the bin content be a Poisson random variable with a mean that is the
same for each bin.

* Benchmark runOnePoint()

Benchmark runOnePoint() so that a fit is actually done and there will be
variation in timing.

At the moment only the NumPy and PyTorch backends are benchmarked in the
test function as there are scaling problems with TensorFlow and the
MXNet optimizer has not been completed.

The source that is used is just repeating values to ensure
reproducibility in the CI. Poisson variations with a fixed seed could
also work, but preliminary testing shows that the NumPy optimizer, which
does not take advantage of auto differentiation, is not always able to
complete the fit at high bin values which would result in a failing test.
Of interest is that PyTorch is able to do so.

* Add benchmark summary

Have the benchmark results be sorted by mean time

* Reset graph and session when using TensorFlow backend

As detailed in Issue #102, the tensor graph and session need to be reset
for each run to have the runs be independent from each other and have
reliable timing.

* Add try except guard to reset backend

If the benchmark fails the assert statement for some reason then the
test will exit immediatley. If this happens the backend will not be
reset to the default. To avoid this, wrap the assert in a try/except so
that upon failure the backend is reset to the default.

* Add test for agreeement on q_mu between backends

* Reset the TensorFlow graph first

Instead of running the computations and then resetting the graph and
session for the following runs, reset the graph first so it will be
directly used. There is a chance that undefined behavior might result
otherwise.

* Move benchmarks to benchmarks directory

This will be used in discriminating tests from benchmarks during the CI

* Make CI always test but only benchmark and deploy on master

To speed up the CI process have the only thing that happens during pull
requests be the testing. Then, when things are brought into master run
the benchmarks, build the docs, and deploy to PyPI.