Add spec-schema validation to spec creation fixtures
Using the spec-schema validation ensures that all specs being tested are
constructed as intended.

Additonally, as pyhf.simplemodels.hepdata_like() constructs a spec and
then calls hfpdf on it, the spec that would be constructed is added
instead. This saves having to have a special case just for calling
hepdata_like making the process more standardized.