Add normal_cdf() to PyTorch backend with a hack
The current stable release of PyTorch, v0.3.1, does not have full
distribution support, and so does not contain the method
torch.distributions.Normal.cdf(). As such, it is copied from the dev
source and implimented by hand here.

c.f.
http://pytorch.org/docs/master/_modules/torch/distributions/normal.html#Normal.cdf

Additonally, PyTorch is expecting things passed in the form of an array
like object, however other backends to not. To guard against this
behavior, if a RuntimeError is raised in pytorch_backend.astensor() then
the input is placed into a list before proceeding through the try/except
block.