Clip all q_mu in runOnePoint
As non-eager (lazy) evaluation backends (i.e., TensorFlow) will not know
how to evaluate a try/except block condition then the clipping should
always be performed to ensure that tensorlib.sqrt() is given reasonable
content.

This requires that the TensorFlow backend .clip() understands how to
handle None types. This is handled by getting the min/max argument of
the input tensor when min/max is None.