Add tensorize decorator
The tensorize decorator maps all the non self arguments of a pyhf.tensor
class method to tensors.

Given that the decorator is used in the tensor module, it must be
declared before the module imported. Ideally this decorator would be
defined in the tensor __init__.py file, but for starters it is being
defined in pyhf.