Install PyTorch with pip (#101)
* Install PyTorch with pip

Now that PyTorch is on PyPI Conda is no longer required to get an easy
install of stable builds. As PyTorch was the only reason Conda was being
installed during CI Miniconda no longer needs to be part of the Travis
YAML file. This should result in a speedup of the CI.
c.f. https://twitter.com/pytorch/status/981198007352705025

As the non-GPU versions are being installed by default to work with
Travis the different versions must be installed with wheels for the time
being. This is done by using conditional checks with python_version in
the setup.py file.

* Don't specifcy wheel URL in setup.py

Try explicitly updating NumPy after install

* Use quite pip install

* Don't require setuptools upgrade

* Use pip install for benchmark stage

* Remove --process-dependency-links

--process-dependency-links is being deprecated and should not be used
* Specify NumPy and pytest min version in setup.py

* Remove apt-get update when no upgrade run