Add hello world example and check to tests (#215)
* Add Hello World notebook

* Unpack runOnePoint output manually to support Python2

As Python2 doesn't have extended unpacking (c.f. PEP 3132) the unpacking
in the example needs to be changed to pass the CI

* Add testing of Hello World example in README

This is done by explicitly calling doctest on the README in the docs
section of the CI. This is done instead of adding the README to
--doctest-glob in pytest.ini so as to keep the example in the README
displaying Python3 code.