Unfreeze dependencies to pass tests
With the current set of frozen depenancies the tests fail due to a
conflict with the coverage library release.

If there is no absolute requirement to freeze a library then it seems
that a lower bound release should be set and then it should be free to
float up as long as it doesn't break anything.