Install NumPy through SciPy's dependency on it
As scipy is an installation requirement then as scipy requires numpy the
latest release of numpy from PyPI will be installed if pyhf is installed
in the default manner. However, if an optional setup is requested
through extra_require and that setup has numpy constraints those
constraints will be implimented.

So

pip install .

installs the latest numpy release

and

pip install .[tensorflow]

installs numpy as constrained by tensorflow.