Set NumPy version in Travis CI for consistency
Travis CI apparently has different behvior for installation then Docker
images such as the official Python 3.6.6 image
(https://hub.docker.com/_/python/) in that the version of NumPy that
results from installation seems to follow a different convention.

Locally, installing with

pip install -U -q -e .[tensorflow,torch,mxnet,develop]

results in a version of numpy that is 1.14.5. However, on Travis CI it
is 1.13.1. To avoid inconsistencies in the doctest and other behavior,
explicitly set the numpy version used to mimic TensorFlow's requirements
but stay above 1.14.0 (which should be done automatically on a local
install).

Trouble that can arise on Travis CI if this is not done, such as getting
this from papermill

RuntimeError: module compiled against API version 0xc but this version
of numpy is 0xb

and the doctest failing.