Remove Docker Hub build status badge (#245)
Given how the repo's Dockerfile is made it will build correctly on
Docker Cloud, but fails on Docker Hub. If the build status badge is left
in the README it will give the false impression that the Docker image is
broken. c.f. Issue #239 for more details on this.

Additionally fix the broken rendering of the uproot link in the RST
version of the README (the website index page)