Revert back to using tf.distribution instead of tfp
For reasons not fully understood, using TensorFlow Probability causes
Travis CI to use too much memory and a core dump occurs.

So until this is understood all use of tfp should be reverted to a
combination of tf.distributions and
tf.contrib.distributions (for the Poisson).

This is really annoying, but at least a temporary fix that doesn't break
all functionality and gains.

c.f.
- https://github.com/pytest-dev/pytest/issues/3527
- https://github.com/travis-ci/travis-ci/issues/9153