Use distributions in ML backends for better Poisson approximation (#268)
* Use TensorFlow Probability for distributions

The only real change is swapping out tf.distributions for
tfp.distributions, however this is a large improvement on the backend.

In addition, tfp is able to better approximate the continous
approximation to the Poisson distribution's p.m.f. while still providing
smooth gradients.

Also add some docstrings.

* Don't use the Normal approximation for the NumPy backend in testing

With improvments to the ML backends it isn't needed

* Use torch.distributions.Poisson for poisson p.m.f. approximation

Provides better approximation then the Normal approximation

* Use PyTorch's Poisson log_prob approximation for MXNet's poisson

This is just implimenting the code that exists at the moment in
PyTorch's library

* Add docstrings to NumPy backend poisson and normal

* Move examples to before Args and Returns

For style consistency in the docs

* Remove use of poisson_from_normal=True from everywhere

It is no longer needed for comparisons and so is not needed anywhere

Though as a result of this, loosen the tolerance for the standard
deviation of results in test_tensor to 5e-5

* Wrap json.load in with clause to safely load and close

Resolves issue of ResourceWarning: unclosed file <_io.TextIOWrapper
name='<your path here>/pyhf/pyhf/data/spec.json' mode='r' encoding='UTF-8'>

Additionally apply autopep8

* Wrap click.open_file in with clause to safely load and close

Resolves issue of ResourceWarning: unclosed file

* Split pytest into two runs given Travis memory constraints

If pytest runs all tests at once then it exceeds the alloted memory for
it by Travis CI. To deal with this, tests/test_notebooks is run by itself
in a second run of pytest. This is a bit strange, but seems to work.

As a result of this and the order that pyest runs the tests, the
docstring tests finish just before the import tests run. This means that
the last backend tested is still in memory, and so the NumPy backend
needs to get set as that is what is needed for the test_import.py

* Revert back to using tf.distribution instead of tfp

For reasons not fully understood, using TensorFlow Probability causes
Travis CI to use too much memory and a core dump occurs.

So until this is understood all use of tfp should be reverted to a
combination of tf.distributions and
tf.contrib.distributions (for the Poisson).

This is really annoying, but at least a temporary fix that doesn't break
all functionality and gains.

c.f.
- https://github.com/pytest-dev/pytest/issues/3527
- https://github.com/travis-ci/travis-ci/issues/9153

* Add names to notebook tests for visual clarity

* Add a universal fix to starting backend

Calling the set default backend before and ensures that the backend gets
reset even if it just came off of a docstring test outside of the tests
dir

As a result, remove the explicit calls to set the NumPy backend in
test_import.py

* Improve docstring with Gamma function comment

Mention that the continuous approximation is done using n! = Gamma(n+1)

Additionally use :code: and :math: qualifiers on different terms to give
specific highlighting in the rendered docs