Pin tensorflow-probability to release v0.3.0
Until a resolution to the issue of tensorflow-probability's change of
the Poisson log_prob in release v0.4.0 to no longer support a continuous
approximation to the Poisson p.m.f. can be reached with the help of the
tfp dev team then tfp should be frozen at v0.3.0 to allow for continued
development. This API change in tfp is breaking and so it can't be
easily fixed without fundamental changes in either pyhf or tfp.

c.f.
- https://github.com/tensorflow/probability/blob/r0.3/tensorflow_probability/python/distributions/poisson.py#L141-L142
- https://github.com/tensorflow/probability/blob/r0.4/tensorflow_probability/python/distributions/poisson.py#L127-L135