fix up indexing of summed_mask in shapesys, staterror (#333)
The main difficulty with shapesys/staterror as it affects multiple (non-consecutive) bins is the way we do masking. For situations where we need to figure out whether to enable the mask or not, it was filling in -1 when the parameter index was 0 which is not good. Additionally, tensorflow.gather() is not happy with -1.