Check _returns length instead of if kwargs is empty
This enforces a consistent return type of the observed CLs. Without it
then

>>> CLs_obs = pyhf.utils.hypotest(...)

would have CLs_obs be of type tuple, whereas

>>> CLs_obs, some_other_return = pyhf.utils.hypotest(..., **kwargs)

has CLs_obs being assigned an element of the returned tuple instead. The
returned object is the same, but the difference in what type of object
the variable is assigned is not desirable.