Change virtual environment link to PyPA page
The PyPA page on Installing Packages has a much more generally helpful
section on Creating Virtual Environments that gives info for Python 2
and Python 3, has relevant links to docs, and is followed up with pip
information. This is a much more useful site to direct new users to then
the virtualenvwrapper docs.