Use h.variances.any() to get truth value
Before this change, tests/test_import.py would give an error of

E       ValueError: The truth value of an array with more than one
element is ambiguous. Use a.any() or a.all()

in extract_error due to a change in the return structure of
uproot_methods TH1.variances return structure in uproot_methods PR 21
where the returned value changed from a list to a numpy.array

* https://github.com/scikit-hep/uproot-methods/pull/21
* https://github.com/scikit-hep/uproot-methods/commit/c2a78e4e05eb710c74d5c6ec5d017f70af28566b#diff-67fbe39413c91feb6ad80acd70f76d47

This change was introduced in release v0.3.0

https://github.com/scikit-hep/uproot-methods/releases/tag/0.3.0

This necessitated a change in the query of h.variances on the pyhf side.