Require PyTorch v1.0 or newer and fix static TLS load error
In requiring the latest version of PyTorch (v1.0) the loading of it
after other imports in tests/conftest.py causes there to be an error
during setup of the pytest fixtures

ImportError while loading conftest
'/home/travis/build/diana-hep/pyhf/tests/conftest.py'.
tests/conftest.py:45: in <module>
    (pyhf.tensor.pytorch_backend(), None),
pyhf/tensor/__init__.py:24: in __getattr__
    e,
E   pyhf.exceptions.ImportBackendError: ('There was a problem importing
PyTorch. The pytorch backend cannot be used.', ImportError('dlopen:
cannot load any more object with static TLS',))

TLS appears to stand for thead local storage

* https://stackoverflow.com/questions/19268293/matlab-error-cannot-open-with-static-tls

and seems to be a reoccuring issue with PyTorch. The advice for how to
deal with it from PyTorch Issue 2083 and 2575 seems to be to simply
change the import order so that

import torch

is always the first import in the files that are causing the issue. The
rational for this fix is not understood to me, other then it just makes
sure that libraries that (presumably) require a large amount of static
TLS are dealt with first.

* https://github.com/pytorch/pytorch/issues/2083
* https://github.com/pytorch/pytorch/issues/2575