Change after_success skip to an echo
Travis CI currenlty has after_success: skip result in everything
following in that stage being skipped. This is not the desired effect
and so if instead of using the skip keyword a throwaway echo is used
then the desired effect of having coveralls not get run will still hold
but still allow for a deployment to occur.