Use PEP 440/508 compatible version syntax in setup.py
To ensure stable APIs for the computational backends while not pinning
at specific versions --- to keep pyhf acting like a proper library and
not an application --- use PEP 440/508's compatible version syntax for
the backend releases to pin them at their current major release version
while allowing for minor version updates. This is also done for pytest,
uproot, and papermill.

This also has the nice feature of being able to intentionally choose
when to update to a newer backend API even if the backend has had a new
major release with API breaking changes for a long time.

c.f.
- https://www.python.org/dev/peps/pep-0440/#compatible-release
- https://www.python.org/dev/peps/pep-0508