Specify down only to minor release
If the release is specified down to the bug release (X.Y.Z) then it will
be pinned down to the minor release (X.Y). c.f. PEP 404:

  the following groups of version clauses are equivalent:
  ~= 1.4.5
  >= 1.4.5, == 1.4.*

so as updates from other libraries are desirable, as long as there is no
API breaking changes, then all dependencies should be specified as

~= X.Y