feat: Add roadmap document for 2019 into 2020
This is a document form of Issue 561 which will allow developers to mark
off in the ROADMAP document how their PR contributes to it.
As it is possible for the checklist in Issue 561 to get out of sync with
developement this should help serve as a source of "truth" about
developement status.