feat: Add setup for custom modifiers (#1625)
* This API breaking change allows for the start of addition of custom
modifiers.
* `pyhf.workspace.Workspace.parameters` are fully removed from the
API through removal from `pyhf.mixins`. Instead, parameters is added
as an attribute to the model config.
* Remove the modifier registry and in place add {modifier}_builder
classes for the modifiers. Additionally add
pyhf.modifiers.histfactory_set which is the map of all the
predefined modifiers that HistFactory specifies.
* All workspace parameters related tests are removed and tests are
added for the change to model.config
* The change from workspace.parameters to model.config.parameters
is made in all affected example notebooks.
* Apply a tiny relative tolerance change in the regression tests to
increase to rel=1.5e-5. The resulting tolerance is on the order of 1e-7
and so the change is inconsequential.