Use try except block to get around tarfile.is_tarfile prior to Python 3.9
tarfile.is_tarfile requires a serialized object prior to Python 3.9, so
there is no way to use a BytesIO object in Python 3.7 and Python 3.8 and
avoid pre-mature serialization. To get around this, just use a try
except block to see if the file is valid as a tarfile or not.