test: Make fail_backend markers add pytest.mark.xfail and remove fail_jax marker on percentile tests (#1730)
* Move `xfail_strict = true` into [tool.pytest.ini_options] to have it
be picked up by pytest properly.
   - c.f. https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
   - Amends PR #1702
* Edit the conftest.py `backend` fixture to mark test nodes that have
`@pytest.mark.fail_{backend}` markers (e.g. fail_jax) with
pytest.mark.xfail to ensure that the test are actually run and checked
to make sure that they don't pass when when they are expected to fail.
This is necessary as using pytest.xfail actually skips the test as it
is assumed that the test is known to fail for reasonable reasons, and
pytest.mark.xfail is intended to be the way to check fail status.
c.f.:
   - https://docs.pytest.org/en/6.2.x/skipping.html#xfail
   - https://docs.pytest.org/en/6.2.x/reference.html#pytest-xfail
   - https://docs.pytest.org/en/6.2.x/reference.html#pytest-mark-xfail-ref
* Remove fail_jax marker from percentile tests and remove only_jax
tests as no longer needed given jax v0.2.26 addresses the problems
they were added to work around.
* Remove fail_pytorch marker from test_ND_gather as the test is now
passing, causing an error given xfail_strict.