ci: Make `build` error on warnings
* Add `PYTHONWARNINGS=error` to the environment to cause Python to treat warnings
as errors during the package build to ensure that the build configuration conforms to
the latest standards. Use the 'PYTHONWARNINGS' environment variable as `build` runs in
a new process and so something like `python -Werror -m build .` won't be able to pass
the flags into the new process, but the shell environment will be picked up.

c.f. https://github.com/pypa/build/ Issue 482