test: Add lower bound on supported numpy versions (#2087)
Set lower bound on numpy to v1.21.0 as this is the lower bound required by JAX v0.4.1
in addition to the lower bound of numpy>=1.14.5 by scipy v1.5.1 (the current lowest
supported version). Newer versions of scipy place an upper bound on numpy for
compatibility reasons and so this is only required for older scipy releases.