Fix if-statement in strategy setting of minuit
The relevant line is supposed to set the strategy of minuit to one of
the supported values 0, 1, 2. Since the strategy `0` evaluates to
`False`, the strategy `0` couldn't be set when creating the optimizer.
The default value in the case of no specified strategy is `None`,
meaning that the statement `is not None` solves the issue.