Modify existing tests to include the fixed case
In the case that no strategy is specified it should default to `not
do_grad`. This is tested in the test `test_minuit_strategy_do_grad`.
In the test `test_minuit_strategy_global` a strategy is specified, when
the minimizer is set. In the case of strategy `0` and no strategy passed
to the `fit` function, however, the strategy defaults again to `not
do_grad`. This doesn't seem to be the desired behaviour as a strategy
was passed in the beginning. Therefore, I changed this assert to fail,
if the final strategy is different from the one assigned in the
beginning.

With this change one test case fails before applying this fix.
The fix in this pull request fixes the test failure.
(The failing case is strategy `0` with the `numpy` backend as
 `default_do_grad` is `False` for this backend.)