Fix dead objects being called through weakrefs
When events (such as 'tensorlib_changed') are called, a bug occured
previously where objects (i.e. the `TensorViewer`) are already dead.
In this case an error occured when member functions where using
attributes such as `self._sorted_indices`.

This bug is fixed by catching the cases, in which the object is
invalidated by earlier callbacks.