Add q0 test statistic as valid option to CLI infer interface
The CLI API (e.g. `pyhf cls`) allows the test statistic to be set,
but it limits the options to `q` and `qtilde`. However, the discovery
test statistic `q0` is also a valid option here, but it can't currently
be set on the command line. This commit is a simple fix; it adds `q0`
to the list of allowed arguments to `pyhf cls --test-stat`.